Extended adaptive regulator to function from dedicated Master core
In the new design the master thread shall execute on the CPU0.
The master thread will hold a logic to predict and
assign the memory bandwidth budgets to the regulated cores.
Physical cores  - CPU1,CPU2, CPU3 and CPU4 will be regulated.

Current issues:
The throttle threads are introduced for each core , which is expected to
waiti/sleep until a signal is triggered to start throttling . However the
wait_event_interruptable() responsible for sleeping  does not wait.
The thread continues to loop wihtout waiting and does not free up the
core.