Master thread throttles and unthrottles the cpu core
The master has the functionality to control throttling of the given
CPU core

Each core has a throttle thread that is created during the loading of the
module. At the outset the thread is waiting and in a TASK_INTERRUPTIBLE
state.The objective of the this thread is to block the core to avoid
sending any LLC read or write request.This is done when the throttle
thread starts looping indefinitely (see ar.c thorttle_thread_func1())
and essentially freezes the all the processes in the core.

The master thread can activate core throttling by
sending a wakeup signal to the throttle thread of a particular core
(using wake_up_interruptible()).It also sets a bit (throttler_task bit , ar.h), that decides
whether the thread continues to loop or not.

Similarly the thread deactivates core throttling by re-setting throttle bit, which makes the
throttle thread go back to sleep.